jb-generic-qa/pyproject.toml
[tool.poetry]
name = "jugalbandi"
version = "0.1.0"
description = ""
authors = ["OpenNyAI Team <opennyai@googlegroups.com>"]
readme = "README.md"
packages = [{include = "generic_qa"}]

[tool.poetry.dependencies]
python = ">=3.10, <4.0.0"
fastapi = "^0.95.2"
pydantic = "^1.10.7"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
google-cloud-storage = "^2.9.0"
python-multipart = "^0.0.6"
faiss-cpu = "^1.7.4"
pypdf2 = "^3.0.1"
pydub = "^0.25.1"
google-cloud-texttospeech = "^2.14.1"
google-cloud-translate = "^3.11.1"
google-cloud-speech = "^2.19.0"
asyncpg = "^0.27.0"
gpt-index = "0.5.25"
langchain = "0.0.142"
boto3 = "^1.26.135"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
httpx = "^0.24.1"
aiofiles = "^23.1.0"
types-aiofiles = "^23.1.0.3"
utils = "^1.0.1"
transitions = "^0.9.0"
wasabi = "^1.1.1"
jb-core = {path = "../packages/jb-core", develop = true}
jb-auth-token = {path = "../packages/jb-auth-token", develop = true}
jb-auth-service = {path = "../jb-auth-service", develop = true}
jb-audio-converter = {path = "../packages/jb-audio-converter", develop = true}
jb-qa = {path = "../packages/jb-qa", develop = true}
jb-translator = {path = "../packages/jb-translator", develop = true}
jb-speech-processor = {path = "../packages/jb-speech-processor", develop = true}
jb-feedback = {path = "../packages/jb-feedback", develop = true}
jb-document-collection = {path = "../packages/jb-document-collection", develop = true}
pymupdf = "1.22.3"
python-docx = "^0.8.11"
docx2txt = "^0.8"
scikit-learn = "^1.2.2"
jb-storage = {path = "../packages/jb-storage", develop = true}
jb-tenant = {path = "../packages/jb-tenant", develop = true}

[tool.pytest.ini_options]
pythonpath = [
  "."
]

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.3.0"
mypy = "^1.3.0"
poethepoet = "^0.20.0"
pytest-asyncio = "^0.21.0"
pytest-mock = "^3.10.0"
gunicorn = "^20.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poe.tasks.lint]
shell = """
black generic_qa tests
flake8 generic_qa tests
mypy generic_qa tests
"""
interpreter = "bash"
help = "format, lint, typecheck"


[tool.poe.tasks.test]
cmd = "python -m pytest -vv -o log_cli=1 -o log_cli_level=INFO -W 'ignore::DeprecationWarning' $FILE"
args = [{name="FILE", default="tests", positional=true}]
help = "run tests using pytest"


[tool.poe.tasks.start]
shell = """
gunicorn -b :8080 --reload -w 2 -k uvicorn.workers.UvicornWorker generic_qa.server:app
"""
help = "start the server for local development"


[tool.poe.tasks.build]
shell = """
docker build --platform=linux/amd64  -f tools/Dockerfile -t gcr.io/indian-legal-bert/jugalbandi_genericqa:latest .. 
"""
help = "build docker container"
